+++ app/src/main/java/com/example/coins/CoinActivity.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
 import com.bumptech.glide.Glide
+import com.example.coins.apiCoin.Coin
 
 class CoinActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
