+++ app/src/main/java/com/example/coins/SplSecondActivity.kt
@@ -1,14 +1,12 @@
 package com.example.coins
 
 import android.content.Intent
-import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import android.util.Log
 import android.view.View
-import android.view.WindowInsets
-import android.view.WindowManager
+import com.example.coins.apiCoin.Coin
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
